Fixes sax-js bug 233 (processing instruction ending with "?")
This is the fix for https://github.com/isaacs/sax-js/issues/233 , but fixed in our fork of the https://github.com/isaacs/sax-js repo since that repo doesn't seem to be getting any attention.  The equivalent PR in that repo is https://github.com/isaacs/sax-js/pull/234 - I've just copied the code here.